feat: implement profile management commands with subcommands for viewing, setting, and removing user information

+++ src/bot/commands/generic/profile.ts
@@ -1,8 +1,23 @@
-import type { Command } from '#bot/types/command';
+import type {
+  Command,
+  CommandExecutors,
+  CommandGroupExecutors,
+  SubcommandExecutor,
+} from '#bot/types/command';
+import type { Component } from '#bot/types/component';
+import type { Modal } from '#bot/types/modal';
+import { sendError } from '#bot/utils/sendError';
 import db from '#db';
-import { users, type User } from '#db/schema';
-import env from '#env';
 import {
+  socialsLinks,
+  users,
+  type User,
+  type UserWithSocials,
+} from '#db/schema';
+import env from '#env';
+import { stripIndents } from 'common-tags';
+import {
+  ActionRowBuilder,
   ApplicationCommandType,
   ButtonStyle,
   ChatInputCommandInteraction,
@@ -10,16 +25,18 @@ import {
   ComponentType,
   ContextMenuCommandBuilder,
   MessageFlags,
+  ModalSubmitInteraction,
   SlashCommandBuilder,
+  TextInputBuilder,
+  TextInputStyle,
   time,
   TimestampStyles,
   UserContextMenuCommandInteraction,
   type User as DiscordUser,
+  type ModalActionRowComponentBuilder,
 } from 'discord.js';
-import { eq } from 'drizzle-orm';
-import type { Component } from '#bot/types/component';
-import { sendError } from '#bot/utils/sendError';
-import { stripIndents } from 'common-tags';
+import { and, eq } from 'drizzle-orm';
+import { z } from 'zod/v4';
 import toUnixTimestamps from '~/utils/toUnixTimestamps';
 
 export const commands = <Command[]>[
@@ -36,18 +53,115 @@ export const commands = <Command[]>[
               .setName('user')
               .setDescription('To view the profile of another user'),
           ),
+      )
+      .addSubcommandGroup((group) =>
+        group
+          .setName('set')
+          .setDescription('Set information in your profile')
+          .addSubcommand((subcommand) =>
+            subcommand.setName('bio').setDescription('Set your profile bio'),
+          )
+          .addSubcommand((subcommand) =>
+            subcommand
+              .setName('nickname')
+              .setDescription('Set your profile nickname')
+              .addStringOption((option) =>
+                option
+                  .setName('nickname')
+                  .setDescription('Your new nickname')
+                  .setMaxLength(32)
+                  .setRequired(true),
+              ),
+          )
+          .addSubcommand((subcommand) =>
+            subcommand
+              .setName('social')
+              .setDescription('Add a social link to your profile')
+              .addStringOption((option) =>
+                option
+                  .setName('name')
+                  .setDescription('The name of the social platform')
+                  .setRequired(true),
+              )
+              .addStringOption((option) =>
+                option
+                  .setName('link')
+                  .setDescription('The link to the social profile')
+                  .setRequired(true),
+              ),
+          ),
+      )
+      .addSubcommandGroup((group) =>
+        group
+          .setName('remove')
+          .setDescription('Remove information from your profile')
+          .addSubcommand((subcommand) =>
+            subcommand.setName('bio').setDescription('Remove your profile bio'),
+          )
+          .addSubcommand((subcommand) =>
+            subcommand
+              .setName('nickname')
+              .setDescription('Remove your profile nickname'),
+          )
+          .addSubcommand((subcommand) =>
+            subcommand
+              .setName('social')
+              .setDescription('Remove a social link from your profile')
+              .addStringOption((option) =>
+                option
+                  .setName('name')
+                  .setDescription('The name of the social platform')
+                  .setRequired(true)
+                  .setAutocomplete(true),
+              ),
+          ),
       ),
     async execute(intr: ChatInputCommandInteraction) {
-      switch (intr.options.getSubcommand()) {
-        case 'view': {
-          const user = intr.options.getUser('user') ?? intr.user;
-          await view(intr, user);
-          break;
-        }
-        default:
-          await sendError(intr, 'Unknown subcommand');
-          break;
+      const groupName = intr.options.getSubcommandGroup(false);
+      const subcommandName = intr.options.getSubcommand();
+
+      let executor: SubcommandExecutor | undefined;
+
+      if (groupName) {
+        const group = executors[groupName];
+        if (group && typeof group === 'object' && !Array.isArray(group))
+          executor = (group as CommandExecutors)[subcommandName];
+      } else {
+        const directExecutor = executors[subcommandName];
+        if (typeof directExecutor === 'function')
+          executor = directExecutor as SubcommandExecutor;
       }
+
+      if (executor) await executor(intr);
+      else await sendError(intr, 'Unknown subcommand');
+    },
+    async autocomplete(intr) {
+      if (
+        intr.options.getSubcommandGroup(false) !== 'remove' ||
+        intr.options.getSubcommand() !== 'social'
+      )
+        return;
+
+      const focused = intr.options.getFocused();
+      const user = await db.query.users.findFirst({
+        columns: { id: true }, // The minimum possible
+        where: eq(users.discord_id, intr.user.id),
+        with: {
+          socials: true,
+        },
+      });
+      if (!user) return await intr.respond([]);
+
+      const filtered = user.socials.filter((social) =>
+        social.name.toLowerCase().includes(focused.toLowerCase()),
+      );
+
+      await intr.respond(
+        filtered.map((social) => ({
+          name: social.name,
+          value: social.name,
+        })),
+      );
     },
   },
   {
@@ -59,11 +173,34 @@ export const commands = <Command[]>[
   },
 ];
 
+const executors: CommandGroupExecutors = {
+  view: async (intr: ChatInputCommandInteraction) => {
+    const user = intr.options.getUser('user') ?? intr.user;
+    await view(intr, user);
+  },
+  set: {
+    bio: setBio,
+    nickname: setNickname,
+    social: addSocial,
+  },
+  remove: {
+    bio: removeBio,
+    nickname: removeNickname,
+    social: removeSocial,
+  },
+};
+
 async function view(intr: CommandInteraction, user: DiscordUser) {
   await intr.deferReply();
 
   const data = await verifyUser(intr, user);
   if (!data) return;
+
+  const socials = await db.query.socialsLinks.findMany({
+    columns: { name: true, url: true },
+    where: eq(socialsLinks.user_id, data.id),
+  });
+  const haveSocials = socials.length > 0;
 
   const registeredAt = toUnixTimestamps(data.created_at.getTime());
 
@@ -85,19 +222,255 @@ async function view(intr: CommandInteraction, user: DiscordUser) {
                 type: ComponentType.TextDisplay,
                 content: stripIndents`
                 # ${user.username}
-                ### **ID**
+                ### ID
                 \`${data.id}\`
-                ### **Registered At (Bot)**
+                ### Registered At
                 ${time(registeredAt, TimestampStyles.ShortDateTime)} (${time(registeredAt, TimestampStyles.RelativeTime)})
-                ### **Nickname**
+                ### Nickname
                 \`${data.nickname ?? 'None'}\`
-                ### **Bio**
+                ### Bio
                 \`\`\`${data.bio ?? 'None'}\`\`\`
-
-                > For information about the user discord profile, use /user-info or 'User Info' context menu command.
                 `,
               },
             ],
+          },
+          {
+            type: ComponentType.TextDisplay,
+            content: '### Socials',
+          },
+          {
+            type: haveSocials
+              ? ComponentType.ActionRow
+              : ComponentType.TextDisplay,
+            content: haveSocials ? '' : '`No social links found`',
+            components: socials.map((social) => ({
+              type: ComponentType.Button,
+              label: social.name,
+              style: ButtonStyle.Link,
+              url: social.url,
+            })),
+          },
+          {
+            type: ComponentType.TextDisplay,
+            content:
+              "> For information about the user discord profile, use /user-info or 'User Info' context menu command",
+          },
+        ],
+      },
+    ],
+  });
+}
+
+async function setBio(intr: ChatInputCommandInteraction) {
+  const data = await verifyUser(intr, intr.user);
+  if (!data) return;
+
+  // In another occasion, a check to ensure the user is the one editing their profile would be needed.
+  // But this command does not accept any user input, so we can skip that.
+  // **But what if Discord is lying to us??? :O**
+
+  // Use modal because the bio can be multiline
+  await intr.showModal({
+    title: 'Edit Bio',
+    customId: 'edit-bio-modal',
+    components: [
+      new ActionRowBuilder<ModalActionRowComponentBuilder>().addComponents(
+        new TextInputBuilder({
+          customId: 'content',
+          label: 'Bio',
+          style: TextInputStyle.Paragraph,
+          placeholder: 'Enter your new bio here',
+          required: true,
+        }),
+      ),
+    ],
+  });
+}
+
+async function setNickname(intr: ChatInputCommandInteraction) {
+  await intr.deferReply({ flags: MessageFlags.Ephemeral });
+
+  const data = await verifyUser(intr, intr.user);
+  if (!data) return;
+
+  // Same stuff as editBio, we can trust the user is the one editing their profile
+
+  const nickname = intr.options.getString('nickname', true)?.trim();
+
+  const updated = await db
+    .update(users)
+    .set({ nickname: nickname })
+    .where(eq(users.id, data.id))
+    .returning({ id: users.id }); // Just to check if successful
+
+  if (updated.length === 0)
+    return await sendError(intr, 'Failed to update your nickname.');
+
+  await intr.followUp({
+    flags: MessageFlags.IsComponentsV2 | MessageFlags.Ephemeral,
+    components: [
+      {
+        type: ComponentType.Container,
+        accent_color: env.EMBED_COLOR,
+        components: [
+          {
+            type: ComponentType.TextDisplay,
+            content: `# Your nickname has been updated!`,
+          },
+        ],
+      },
+    ],
+  });
+}
+
+async function addSocial(intr: ChatInputCommandInteraction) {
+  await intr.deferReply({ flags: MessageFlags.Ephemeral });
+
+  const data = await verifyUser(intr, intr.user);
+  if (!data) return;
+
+  const socials = await db.query.socialsLinks.findMany({
+    columns: { id: true, name: true }, // The minimum possible
+    where: eq(socialsLinks.user_id, data.id),
+  });
+
+  if (socials.length >= 5)
+    return await sendError(
+      intr,
+      'You can only have up to 5 social links in your profile.',
+    );
+
+  // Same stuff as editBio, we can trust the user is the one editing their profile
+
+  const name = intr.options.getString('name', true).trim();
+  const link = intr.options.getString('link', true).trim();
+
+  if (
+    socials.some((social) => social.name.toLowerCase() === name.toLowerCase())
+  )
+    return await sendError(
+      intr,
+      `You already have a social link with the name \`${name}\`. Please choose a different name.`,
+    );
+
+  if (!z.url().safeParse(link).success)
+    return await sendError(intr, 'The provided link is not a valid URL.');
+
+  const inserted = await db
+    .insert(socialsLinks)
+    .values({ name, url: link, user_id: data.id })
+    .returning({ id: socialsLinks.id }); // Just to check if successful
+
+  if (inserted.length === 0)
+    return await sendError(intr, 'Failed to add the social link.');
+
+  await intr.followUp({
+    flags: MessageFlags.IsComponentsV2 | MessageFlags.Ephemeral,
+    components: [
+      {
+        type: ComponentType.Container,
+        accent_color: env.EMBED_COLOR,
+        components: [
+          {
+            type: ComponentType.TextDisplay,
+            content: `# The social link has been added!`,
+          },
+        ],
+      },
+    ],
+  });
+}
+
+async function removeBio(intr: ChatInputCommandInteraction) {
+  await intr.deferReply({ flags: MessageFlags.Ephemeral });
+
+  const data = await verifyUser(intr, intr.user);
+  if (!data) return;
+
+  const updated = await db
+    .update(users)
+    .set({ bio: null })
+    .where(eq(users.id, data.id))
+    .returning({ id: users.id }); // Just to check if successful
+
+  if (updated.length === 0)
+    return await sendError(intr, 'Failed to remove your bio.');
+
+  await intr.followUp({
+    flags: MessageFlags.IsComponentsV2 | MessageFlags.Ephemeral,
+    components: [
+      {
+        type: ComponentType.Container,
+        accent_color: env.EMBED_COLOR,
+        components: [
+          {
+            type: ComponentType.TextDisplay,
+            content: `# Your bio has been removed!`,
+          },
+        ],
+      },
+    ],
+  });
+}
+
+async function removeNickname(intr: ChatInputCommandInteraction) {
+  await intr.deferReply({ flags: MessageFlags.Ephemeral });
+
+  const data = await verifyUser(intr, intr.user);
+  if (!data) return;
+
+  const updated = await db
+    .update(users)
+    .set({ nickname: null })
+    .where(eq(users.id, data.id))
+    .returning({ id: users.id }); // Just to check if successful
+
+  if (updated.length === 0)
+    return await sendError(intr, 'Failed to remove your nickname.');
+
+  await intr.followUp({
+    flags: MessageFlags.IsComponentsV2 | MessageFlags.Ephemeral,
+    components: [
+      {
+        type: ComponentType.Container,
+        accent_color: env.EMBED_COLOR,
+        components: [
+          {
+            type: ComponentType.TextDisplay,
+            content: `# Your nickname has been removed!`,
+          },
+        ],
+      },
+    ],
+  });
+}
+
+async function removeSocial(intr: ChatInputCommandInteraction) {
+  await intr.deferReply({ flags: MessageFlags.Ephemeral });
+
+  const data = await verifyUser(intr, intr.user);
+  if (!data) return;
+
+  const name = intr.options.getString('name', true).trim();
+
+  const deleted = await db
+    .delete(socialsLinks)
+    .where(and(eq(socialsLinks.user_id, data.id), eq(socialsLinks.name, name)))
+    .returning({ id: socialsLinks.id }); // Just to check if successful
+
+  if (deleted.length === 0)
+    return await sendError(intr, 'Failed to remove the social link.');
+
+  await intr.followUp({
+    flags: MessageFlags.IsComponentsV2 | MessageFlags.Ephemeral,
+    components: [
+      {
+        type: ComponentType.Container,
+        accent_color: env.EMBED_COLOR,
+        components: [
+          {
+            type: ComponentType.TextDisplay,
+            content: `# The social link has been removed!`,
           },
         ],
       },
@@ -106,27 +479,23 @@ async function view(intr: CommandInteraction, user: DiscordUser) {
 }
 
 async function verifyUser(
-  intr: CommandInteraction,
+  intr: CommandInteraction | ModalSubmitInteraction,
   user: DiscordUser,
-): Promise<User | void> {
+): Promise<User | UserWithSocials | void> {
   const data = await db.query.users.findFirst({
     where: eq(users.discord_id, user.id),
   });
   if (data) return data;
 
-  if (intr.user.id !== user.id)
-    return await sendError(
-      intr,
-      `The user \`${user.username}\` does not have a profile!`,
-    );
-
-  if (!intr.ephemeral) {
+  if (!intr.ephemeral && intr.replied) {
     // Delete the initial reply because we need it to be ephemeral
     const msg = await intr.followUp('opps...');
     await msg.delete().catch(() => null);
   }
 
-  await intr.followUp({
+  const action = intr.replied ? intr.followUp : intr.reply;
+
+  await action({
     flags: MessageFlags.IsComponentsV2 | MessageFlags.Ephemeral,
     components: [
       {
@@ -142,7 +511,7 @@ async function verifyUser(
             components: [
               {
                 type: ComponentType.Button,
-                custom_id: 'register-profile',
+                custom_id: 'register-profile-button',
                 label: 'Register',
                 style: ButtonStyle.Primary,
                 emoji: { name: '📝' },
@@ -157,7 +526,7 @@ async function verifyUser(
 
 export const component = <Component>{
   type: ComponentType.Button,
-  customId: 'register-profile',
+  customId: 'register-profile-button',
   async execute(intr) {
     await intr.deferReply({ flags: MessageFlags.Ephemeral });
 
@@ -166,7 +535,7 @@ export const component = <Component>{
       .values({
         discord_id: intr.user.id,
       })
-      .returning({ id: users.id })
+      .returning({ id: users.id }) // Just to check if successful
       .onConflictDoNothing();
     if (user.length === 0)
       return await sendError(intr, 'You already have a profile!');
@@ -181,6 +550,45 @@ export const component = <Component>{
             {
               type: ComponentType.TextDisplay,
               content: `# Your profile has been created!`,
+            },
+          ],
+        },
+      ],
+    });
+  },
+};
+
+export const modal = <Modal>{
+  customId: 'edit-bio-modal',
+  async execute(intr) {
+    await intr.deferReply({ flags: MessageFlags.Ephemeral });
+
+    const content = intr.fields.getTextInputValue('content').trim();
+
+    // Ideally, we would check if the user exists in the database,
+    // but this modal is only shown after the user has been verified,
+    // so let's trust that the user has a profile
+    // Doing this we avoid an extra database query
+
+    const updated = await db
+      .update(users)
+      .set({ bio: content })
+      .where(eq(users.discord_id, intr.user.id))
+      .returning({ id: users.id }); // Just to check if successful
+
+    if (updated.length === 0)
+      return await sendError(intr, 'Failed to update your bio.');
+
+    await intr.followUp({
+      flags: MessageFlags.IsComponentsV2 | MessageFlags.Ephemeral,
+      components: [
+        {
+          type: ComponentType.Container,
+          accent_color: env.EMBED_COLOR,
+          components: [
+            {
+              type: ComponentType.TextDisplay,
+              content: `# Your bio has been updated!`,
             },
           ],
         },
